feat: adiciona controllers, servicos, repositorios, dtos e entidades

+++ src/main/java/br/com/fcxlabs/hermes/repositories/SequentialRepository.java
@@ -3,5 +3,5 @@ package br.com.fcxlabs.hermes.repositories;
 import br.com.fcxlabs.hermes.entities.Sequential;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SequentialRepository extends JpaRepository<Sequential, Integer> {
+public interface SequentialRepository extends JpaRepository<Sequential, Long> {
 }
