+++ src/main/java/br/com/fcxlabs/hermes/repositories/TrackerRepository.java
@@ -3,5 +3,5 @@ package br.com.fcxlabs.hermes.repositories;
 import br.com.fcxlabs.hermes.entities.Tracker;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrackerRepository extends JpaRepository<Tracker, Integer> {
+public interface TrackerRepository extends JpaRepository<Tracker, Long> {
 }
